% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation.QoLSEM.R
\name{generation.QoLSEM}
\alias{generation.QoLSEM}
\title{Generation dataset from a QoLSEM model}
\usage{
generation.QoLSEM(N = 500, I = 150, c = as.matrix(c(2, -2)),
  a1 = c(1:7), a2 = c(1:12), d = as.matrix(c(80)),
  D1 = matrix(seq(0:14, 2) + 70, ncol = 7), D2 = matrix(seq(0:24, 2) +
  20, ncol = 12), sigma.y = 10, sigma.X1 = 10, sigma.X2 = 10)
}
\arguments{
\item{N}{number of dataset to generate}

\item{I}{number of observations (statistic units) by datasets; a pair number are expected}

\item{c}{a numerical vector denoting the two scalars c=[c1,c2]: parameters associated with the two factors to explain the response varaible Y}

\item{a1}{a numerical vector (of size q_1) of parameters associated with the factor f_1}

\item{a2}{a numerical vector (of size q_2) of parameters associated with the factor f_2}

\item{d}{vector of explanatory variables associated with y}

\item{D1}{matrix of parameters associated with the block 1}

\item{D2}{matrix of parameters associated with the block 2}

\item{sigma.y}{standard deviation of the response variable y}

\item{sigma.X1}{average standard deviation of the block 1}

\item{sigma.X2}{average standard deviation of the block 2}
}
\value{
A dataframe for N>1 and a list for N=1
}
\description{
This function is used to generate dataset(s) from the model proposed by Barbieri, Tami et al. (2018).
It is used to propose an example to the users.
}
\examples{
## test avec N=1
test1 <- generation.QoLSEM(N=1,I=150,
                           c=as.matrix(c(2,-2)),
                           a1=matrix(c(1:7),nrow=7),
                           a2=matrix(c(1:12),nrow=12),
                           d=c(80),
                           D1=matrix(seq(2,14,2)+70,ncol=7),
                           D2=matrix(seq(2,24,2)+20,ncol=12),
                           sigma.y=10,sigma.X1=10,sigma.X2=10)

## test avec N>1
test20 <- generation.QoLSEM(N=20,I=150,
                            c=as.matrix(c(2,-2)),
                            a1=matrix(c(1:7),nrow=7),
                            a2=matrix(c(1:12),nrow=12),
                            d=c(80),
                            D1=matrix(seq(2,14,2)+70,ncol=7),
                            D2=matrix(seq(2,24,2)+20,ncol=12),
                            sigma.y=10,sigma.X1=7,sigma.X2=5)

}
\references{
Barbieri A, Tami M, Bry X, Azria D, Gourgou S, Mollevi C, Lavergne C. (2018) \emph{EM algorithm estimation of a structural equation model for the longitudinal study of the quality of life}. Statistics in Medicine. 37(6) :1031-1046.
}
\author{
Antoine Barbieri, Myriam Tami
}
